Rename figures-counter.* to follow convention with '_' instead '-'
