+++ figures-counter/includes/figures_counter.h
